+++ common/src/main/java/net/mehvahdjukaar/sleep_tight/mixins/LivingEntityMixin.java
@@ -28,7 +28,7 @@ public abstract class LivingEntityMixin extends Entity {
     }
 
     @Inject(method = "setPosToBed", at = @At("HEAD"), cancellable = true)
-    public void setHammockPos(BlockPos pos, CallbackInfo ci) {
+    public void sleep_tight$setHammockPos(BlockPos pos, CallbackInfo ci) {
         BlockState state = this.level().getBlockState(pos);
         Vec3 v = ModEvents.getSleepingPosition(this, state, pos);
         if (v != null) {
@@ -37,7 +37,7 @@ public abstract class LivingEntityMixin extends Entity {
         }
     }
     @Inject(method = "isSleeping", at = @At(value = "HEAD"), cancellable = true)
-    public void sleepOnEntity(CallbackInfoReturnable<Boolean> cir) {
+    public void sleep_tight$sleepOnEntity(CallbackInfoReturnable<Boolean> cir) {
         if (this.level().isClientSide && !this.isDeadOrDying() && this.getVehicle() instanceof BedEntity && ClientEvents.cameraHack) {
             cir.setReturnValue(true);
         }